Move ValidateWebhook from Hook to Github
