+++ prow/external-plugins/cherrypicker/BUILD.bazel
@@ -13,7 +13,6 @@ go_library(
         "//prow/flagutil:go_default_library",
         "//prow/git:go_default_library",
         "//prow/github:go_default_library",
-        "//prow/hook:go_default_library",
         "//prow/pluginhelp:go_default_library",
         "//prow/pluginhelp/externalplugins:go_default_library",
         "//prow/plugins:go_default_library",
