+++ prow/external-plugins/cherrypicker/server.go
@@ -32,7 +32,6 @@ import (
 
 	"k8s.io/test-infra/prow/git"
 	"k8s.io/test-infra/prow/github"
-	"k8s.io/test-infra/prow/hook"
 	"k8s.io/test-infra/prow/pluginhelp"
 	"k8s.io/test-infra/prow/plugins"
 )
@@ -96,7 +95,7 @@ type Server struct {
 
 // ServeHTTP validates an incoming webhook and puts it into the event channel.
 func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	eventType, eventGUID, payload, ok := hook.ValidateWebhook(w, r, s.tokenGenerator())
+	eventType, eventGUID, payload, ok := github.ValidateWebhook(w, r, s.tokenGenerator())
 	if !ok {
 		return
 	}
