+++ prow/external-plugins/needs-rebase/main.go
@@ -33,7 +33,6 @@ import (
 	"k8s.io/test-infra/prow/external-plugins/needs-rebase/plugin"
 	"k8s.io/test-infra/prow/flagutil"
 	"k8s.io/test-infra/prow/github"
-	"k8s.io/test-infra/prow/hook"
 	"k8s.io/test-infra/prow/pluginhelp/externalplugins"
 	"k8s.io/test-infra/prow/plugins"
 )
@@ -111,9 +110,7 @@ type Server struct {
 
 // ServeHTTP validates an incoming webhook and puts it into the event channel.
 func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
-	// TODO: Move webhook handling logic out of hook binary so that we don't have to import all
-	// plugins just to validate the webhook.
-	eventType, eventGUID, payload, ok := hook.ValidateWebhook(w, r, s.tokenGenerator())
+	eventType, eventGUID, payload, ok := github.ValidateWebhook(w, r, s.tokenGenerator())
 	if !ok {
 		return
 	}
