+++ prow/github/BUILD.bazel
@@ -13,6 +13,7 @@ go_test(
         "hmac_test.go",
         "links_test.go",
         "types_test.go",
+        "webhooks_test.go",
     ],
     embed = [":go_default_library"],
     deps = ["//vendor/k8s.io/apimachinery/pkg/util/sets:go_default_library"],
@@ -26,6 +27,7 @@ go_library(
         "hmac.go",
         "links.go",
         "types.go",
+        "webhooks.go",
     ],
     importpath = "k8s.io/test-infra/prow/github",
     deps = [
